search function done

+++ app/src/main/java/com/example/weatherapp/data/remote/model/currentweather/Coord.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Coord(
     @SerializedName("lat")
-    var lat: Int?,
+    var lat: Double?,
     @SerializedName("lon")
-    var lon: Int?
+    var lon: Double?
 )
